Domain Layer Architecture for UC:Process Sale rearrage database connection rename 'stock' table field from 'amount' to 'quantity' quantity is double, not int id use int is enoungh, long is too big

+++ src/com/refresh/pos/database/Database.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 public interface Database {
 	public List<Object> select(String queryString);
-	public long insert(String tableName, Object content);
+	public int insert(String tableName, Object content);
 	public boolean update();
 	public boolean delete();
 }
